Spinlock wait for process
I noticed the new processes were getting started too quickly and
not allowing the killed ones to terminate gracefully.  This
waits for the SIGTERM process to stop before proceeding with the
restart.  Ideally this should wait a certain period of time and
then SIGKILL the process but I'm not there yet.